refactor: update component imports to use shadcn directory
